yt_app/requirements.txt
flask==2.3.3
werkzeug==2.3.7
pytube==15.0.0
youtube-search==2.1.0
ffmpeg-python==0.2.0
flask-cors==4.0.0
retry==0.9.2
